Restructure project base on golang-standards/project-layout
