+++ cmd/main.go
@@ -1,0 +1,12 @@
+// #cgo CFLAGS: -O3
+// #cgo CXXFLAGS: -O3
+package main
+
+import (
+	"github.com/andlabs/ui"
+	"github.com/limingjie/BarbarasTool/cmd/gui"
+)
+
+func main() {
+	ui.Main(gui.NewGUI)
+}
